MLOPs/hacker_news/requirements.txt
prefect
beautifulsoup4
httpx
jupyter
psycopg2-binary
